Apps and tests default to assuming XCORE-AI-EXPLORER.
This changes each test and example to default XCORE_TARGET to XCORE-AI-EXPLORER if the user hasn't configured the CMake project with the XCORE_TARGET variable.

This way users can use the CMAKE_TOOLCHAIN_FILE located in the xcore_sdk without needing to provide a target, as they could before the previous commit.